Replace detailed collision meshes with convex hulls
The original STL files used for collision meshes for the Robotiq 85 gripper
were high-resolution detailed meshes similar to the ones used for visual models.
These caused problems when loaded into a planning environment in some cases,
for example causing the environment to take a long time to load or update.
In any case it isn't necessary to use detailed meshes to represent collision
geometry, so this changes replaces them with lighter-weight convex hull meshes.